zenvoco-backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-jose
passlib
bcrypt
python-multipart
openai
google-generativeai==0.8.6
assemblyai==0.33.0
librosa==0.10.2
numpy==2.0.2
